requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
click==8.1.7
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.10.0
PyYAML==6.0.2
referencing==0.35.1
rest-framework-simplejwt==0.0.2
rpds-py==0.21.0
sqlparse==0.5.2
uritemplate==4.1.1
